chore(web): move `grid.util` inside `grid/`
