+++ packages/web/src/common/utils/draft/draft.util.ts
@@ -9,7 +9,7 @@ import {
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import { roundToNext } from "@web/common/utils";
 import { assembleDefaultEvent } from "@web/common/utils/event/event.util";
-import { getElemById, getX } from "@web/common/utils/grid.util";
+import { getElemById, getX } from "@web/common/utils/grid/grid.util";
 import { draftSlice } from "@web/ducks/events/slices/draft.slice";
 import { Activity_DraftEvent } from "@web/ducks/events/slices/draft.slice.types";
 import { DateCalcs } from "@web/views/Calendar/hooks/grid/useDateCalcs";
