+++ packages/web/src/views/Calendar/components/Draft/grid/hooks/useGridMouseUp.ts
@@ -1,6 +1,6 @@
 import { MouseEvent, useCallback } from "react";
 import { Categories_Event } from "@core/types/event.types";
-import { getElemById } from "@web/common/utils/grid.util";
+import { getElemById } from "@web/common/utils/grid/grid.util";
 import { selectDraftStatus } from "@web/ducks/events/selectors/draft.selectors";
 import { useAppSelector } from "@web/store/store.hooks";
 import { useEventListener } from "@web/views/Calendar/hooks/mouse/useEventListener";
