+++ packages/web/src/views/Calendar/components/Event/Grid/GridEvent/GridEvent.tsx
@@ -14,7 +14,7 @@ import {
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import { getTimesLabel } from "@web/common/utils/datetime/web.date.util";
 import { isOptimisticEvent } from "@web/common/utils/event/event.util";
-import { getLineClamp } from "@web/common/utils/grid.util";
+import { getLineClamp } from "@web/common/utils/grid/grid.util";
 import { isRightClick } from "@web/common/utils/mouse/mouse.util";
 import { getEventPosition } from "@web/common/utils/position/position.util";
 import { Flex } from "@web/components/Flex";
