+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/AllDayRow.tsx
@@ -8,7 +8,7 @@ import {
 } from "@web/common/constants/web.constants";
 import { Ref_Callback } from "@web/common/types/util.types";
 import { assembleDefaultEvent } from "@web/common/utils/event/event.util";
-import { getX } from "@web/common/utils/grid.util";
+import { getX } from "@web/common/utils/grid/grid.util";
 import { isRightClick } from "@web/common/utils/mouse/mouse.util";
 import { selectIsDrafting } from "@web/ducks/events/selectors/draft.selectors";
 import { selectRowCount } from "@web/ducks/events/selectors/event.selectors";
