+++ packages/web/src/views/Calendar/hooks/grid/useGridEventMouseDown.test.ts
@@ -2,7 +2,7 @@ import { renderHook } from "@testing-library/react";
 import { Categories_Event } from "@core/types/event.types";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import * as utils from "@web/common/utils";
-import * as gridUtils from "@web/common/utils/grid.util";
+import * as gridUtils from "@web/common/utils/grid/grid.util";
 import {
   GRID_EVENT_MOUSE_HOLD_DELAY,
   GRID_EVENT_MOUSE_HOLD_MOVE_THRESHOLD,
