+++ packages/web/src/views/Calendar/hooks/grid/useScroll.ts
@@ -1,5 +1,5 @@
 import { MutableRefObject, useCallback, useEffect } from "react";
-import { getCurrentMinute } from "@web/common/utils/grid.util";
+import { getCurrentMinute } from "@web/common/utils/grid/grid.util";
 
 export const useScroll = (
   timedGridRef: MutableRefObject<HTMLDivElement | null>,
